move activity styles to own module
